requirements.txt
simpleaudio
requests
pyyaml